refactor: move shared packages up a level
