+++ scrobble/listenbrainz/listenbrainz.go
@@ -10,8 +10,8 @@ import (
 	"net/http/httputil"
 	"time"
 
-	"go.senan.xyz/gonic/server/db"
-	"go.senan.xyz/gonic/server/scrobble"
+	"go.senan.xyz/gonic/db"
+	"go.senan.xyz/gonic/scrobble"
 )
 
 const (
