+++ server/ctrladmin/handlers_playlist.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/jinzhu/gorm"
 
-	"go.senan.xyz/gonic/server/db"
+	"go.senan.xyz/gonic/db"
 )
 
 var (
