+++ server/ctrladmin/middleware.go
@@ -8,7 +8,7 @@ import (
 	"github.com/gorilla/sessions"
 
 	"go.senan.xyz/gonic"
-	"go.senan.xyz/gonic/server/db"
+	"go.senan.xyz/gonic/db"
 )
 
 func (c *Controller) WithSession(next http.Handler) http.Handler {
