+++ server/ctrlbase/ctrl.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 	"path"
 
-	"go.senan.xyz/gonic/server/db"
-	"go.senan.xyz/gonic/server/scanner"
+	"go.senan.xyz/gonic/db"
+	"go.senan.xyz/gonic/scanner"
 )
 
 type statusWriter struct {
