+++ server/ctrlsubsonic/ctrl.go
@@ -12,10 +12,10 @@ import (
 	"go.senan.xyz/gonic/server/ctrlbase"
 	"go.senan.xyz/gonic/server/ctrlsubsonic/params"
 	"go.senan.xyz/gonic/server/ctrlsubsonic/spec"
-	"go.senan.xyz/gonic/server/jukebox"
-	"go.senan.xyz/gonic/server/podcasts"
-	"go.senan.xyz/gonic/server/scrobble"
-	"go.senan.xyz/gonic/server/transcode"
+	"go.senan.xyz/gonic/jukebox"
+	"go.senan.xyz/gonic/podcasts"
+	"go.senan.xyz/gonic/scrobble"
+	"go.senan.xyz/gonic/transcode"
 )
 
 type CtxKey int
