+++ server/ctrlsubsonic/handlers_bookmark.go
@@ -9,7 +9,7 @@ import (
 	"go.senan.xyz/gonic/server/ctrlsubsonic/params"
 	"go.senan.xyz/gonic/server/ctrlsubsonic/spec"
 	"go.senan.xyz/gonic/server/ctrlsubsonic/specid"
-	"go.senan.xyz/gonic/server/db"
+	"go.senan.xyz/gonic/db"
 )
 
 func (c *Controller) ServeGetBookmarks(r *http.Request) *spec.Response {
