+++ server/ctrlsubsonic/handlers_by_tags.go
@@ -13,8 +13,8 @@ import (
 	"go.senan.xyz/gonic/server/ctrlsubsonic/params"
 	"go.senan.xyz/gonic/server/ctrlsubsonic/spec"
 	"go.senan.xyz/gonic/server/ctrlsubsonic/specid"
-	"go.senan.xyz/gonic/server/db"
-	"go.senan.xyz/gonic/server/scrobble/lastfm"
+	"go.senan.xyz/gonic/db"
+	"go.senan.xyz/gonic/scrobble/lastfm"
 )
 
 func (c *Controller) ServeGetArtists(r *http.Request) *spec.Response {
