+++ server/ctrlsubsonic/handlers_podcast.go
@@ -8,7 +8,7 @@ import (
 	"go.senan.xyz/gonic/server/ctrlsubsonic/params"
 	"go.senan.xyz/gonic/server/ctrlsubsonic/spec"
 	"go.senan.xyz/gonic/server/ctrlsubsonic/specid"
-	"go.senan.xyz/gonic/server/db"
+	"go.senan.xyz/gonic/db"
 )
 
 func (c *Controller) ServeGetPodcasts(r *http.Request) *spec.Response {
