+++ server/server.go
@@ -15,15 +15,15 @@ import (
 	"go.senan.xyz/gonic/server/ctrladmin"
 	"go.senan.xyz/gonic/server/ctrlbase"
 	"go.senan.xyz/gonic/server/ctrlsubsonic"
-	"go.senan.xyz/gonic/server/db"
-	"go.senan.xyz/gonic/server/jukebox"
-	"go.senan.xyz/gonic/server/podcasts"
-	"go.senan.xyz/gonic/server/scanner"
-	"go.senan.xyz/gonic/server/scanner/tags"
-	"go.senan.xyz/gonic/server/scrobble"
-	"go.senan.xyz/gonic/server/scrobble/lastfm"
-	"go.senan.xyz/gonic/server/scrobble/listenbrainz"
-	"go.senan.xyz/gonic/server/transcode"
+	"go.senan.xyz/gonic/db"
+	"go.senan.xyz/gonic/jukebox"
+	"go.senan.xyz/gonic/podcasts"
+	"go.senan.xyz/gonic/scanner"
+	"go.senan.xyz/gonic/scanner/tags"
+	"go.senan.xyz/gonic/scrobble"
+	"go.senan.xyz/gonic/scrobble/lastfm"
+	"go.senan.xyz/gonic/scrobble/listenbrainz"
+	"go.senan.xyz/gonic/transcode"
 )
 
 type Options struct {
